Adding local traces
